+act #18913 moved PrettyDuration to akka-actor
